(Total) Final
Removed an if-else that was used for testing and would set an "eid" (primary key) if
no "eid" would be give by the user/$_POST['eid']. Since this isn't needed anymore and
it is safer without so no one can access it without "eid", this part was removed.

Without an "eid", the pdfgenerator now returns a pdf without database-content and
only with pdf-structure and some text.